Create slides
This so far displays a slideshow with the labels and slide name. Next I will work on the play/pause button for the slideshow and will also create a separate file, under the slides tab which will allow the web user to delete and upload slides as well as rearrange them.